backend/requirements.txt
Django==4.2.7
psycopg2-binary==2.9.11
django-cors-headers==4.3.1
python-dotenv==1.0.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
requests==2.31.0
Pillow==12.0.0
djangorestframework==3.14.0
